feat(menu): suppress MenuPdf and add backmenu test
